+++ patientMatcher/server/views.py
@@ -6,6 +6,7 @@ import os
 from bson import json_util
 from flask import Blueprint, current_app, jsonify, render_template, request, send_from_directory
 from flask_negotiate import consumes, produces
+
 from patientMatcher.auth.auth import authorize
 from patientMatcher.match.handler import internal_matcher, patient_matches
 from patientMatcher.utils.add import backend_add_patient